+++ packages/data-schema/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @aws-amplify/data-schema
 
+## 0.13.15
+
+### Patch Changes
+
+- f33a8e6: remove underscores from generated fn names
+
+## 0.13.14
+
+### Patch Changes
+
+- 0171ad5: handler.function supprts string | defineFunction
+- Updated dependencies [0171ad5]
+  - @aws-amplify/data-schema-types@0.7.8
+
+## 0.13.13
+
+### Patch Changes
+
+- 0e36d74: fix: delete operation of m:m created implicit model not having correct param
+
+## 0.13.12
+
+### Patch Changes
+
+- ab4bc27: feat: Schema transform logic for inlineSql and sqlReference handlers
+
 ## 0.13.11
 
 ### Patch Changes